+++ health-services/resource-generator/src/main/java/org/egov/processor/service/FileParser.java
@@ -1,0 +1,13 @@
+package org.egov.processor.service;
+
+
+import java.math.BigDecimal;
+import org.egov.processor.web.models.Plan;
+import org.egov.processor.web.models.PlanConfiguration;
+import org.egov.processor.web.models.PlanConfigurationRequest;
+
+public interface FileParser {
+
+    Object parseFileData(PlanConfigurationRequest planConfigurationRequest , String fileStoreId, Object campaignResponse);
+
+   }