+++ health-services/household/CHANGELOG.md
@@ -1,5 +1,51 @@
 All notable changes to this module will be documented in this file.
 
+## 1.1.7 - 2025-04-11
+
+- Upgraded health-services-models to 1.0.27
+- Added relationship type support for household members
+
+## 1.1.6 - 2025-02-28
+
+- Upgraded to health commons library version 1.0.21
+
+## 1.1.5 - 2025-01-28
+
+- Added householdType column in household table
+- Upgraded to heath models 1.0.25
+
+## 1.1.4 - 2024-08-29 
+
+- Added `ExistentEntityValidator` fixes
+
+## 1.1.3 - 2024-05-29 
+
+- Integrated Core 2.9 LTS
+- Client reference ID validation added
+- Upgraded to health models 1.0.20 and health common 1.0.16
+- Boundary v2 Integration
+- MDMS v2 integration
+- Upgraded PostgresSQL Driver version to 42.7.1
+- Upgraded Flyway base image version to 10.7.1 for DB Migration
+- Upgraded Flyway-Core to 9.22.3
+
+## 1.1.2 - 2024-05-10
+
+- Integrated Boundary v2 functionality
+
+## 1.1.1 - 2023-11-15
+
+- Added total count for household
+- Added a field for HouseholdMember-clientReferenceId
+
+## 1.1.1-beta
+
+- Added proximity based search support
+
+## 1.1.0
+
+
 ## 1.0.0
 
 - Base version
+